requirements.txt
flask>=3.0.0
pandas>=2.2.0
xhtml2pdf>=0.2.11
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
